+++ molo/core/cookiecutter/scaffold/{{cookiecutter.directory}}/{{cookiecutter.app_name}}/settings/base.py
@@ -129,7 +129,7 @@ WSGI_APPLICATION = '{{cookiecutter.app_name}}.wsgi.application'
 # Google analytics
 
 GOOGLE_ANALYTICS = {
-    'google_analytics_id': 'xxx',
+    'google_analytics_id': None,
 }
 
 GOOGLE_ANALYTICS_IGNORE_PATH = [
@@ -328,3 +328,8 @@ EMAIL_HOST = environ.get('EMAIL_HOST', 'localhost')
 EMAIL_PORT = environ.get('EMAIL_PORT', 25)
 EMAIL_HOST_USER = environ.get('EMAIL_HOST_USER', '')
 EMAIL_HOST_PASSWORD = environ.get('EMAIL_HOST_PASSWORD', '')
+
+CSRF_FAILURE_VIEW = 'molo.core.views.csrf_failure'
+
+FREE_BASICS_URL_FOR_CSRF_MESSAGE = environ.get(
+    'FREE_BASICS_URL_FOR_CSRF_MESSAGE', '')