+++ CHANGES.rst
@@ -16,6 +16,10 @@ Main features::
 - Upgraded wagtail to 1.4.3
 - We've added the option to allow un-translated pages to be hidden
 - We now show a translated page on the front end when it's main language page is unpublished
+- Add Topic of the Day functionality
+- Add Support for both Elastichsearch 1.x & 2.x
+- Add ability to show a highlighted term in the results
+- Implement custom error page for CSRF error
 
 Backwards incompatible changes
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
@@ -25,6 +29,54 @@ Backwards incompatible changes
 - Deprecated use of ``Section.featured_articles_in_homepage``: use the template tag ``{% load_descendant_articles_for_section section featured_in_homepage=True %}``
 - Deprecated use of ``Section.latest_articles_in_homepage``: use the template tag ``{% load_descendant_articles_for_section section featured_in_latest=True %}``
 - Deprecated use of ``Section.articles``: use the template tag ``{% load_child_articles_for_section page %}``
+
+3.12.1
+~~~~~~
+
+- (bug) search URL was defined using the wrong regex (it broke Service Directory plugin)
+
+3.12.0
+~~~~~
+
+- Implement custom error page for CSRF error
+
+3.11.2
+~~~~~
+
+- Remove automatic opening of comments when an article is promoted to Topic of the Day
+
+3.11.1
+~~~~~
+
+- Exclude future-dated Topic of the Day articles from Latest articles list
+
+3.11.0
+~~~~~
+
+- Add Support for both Elastichsearch 1.x & 2.x
+- Add ability to show a highlighted term in the results
+Note: Search highlighting is only supported by the Elasticsearch backend.
+
+3.10.0
+~~~~~
+
+- Add Topic of the Day functionality
+
+3.9.2
+~~~~~
+
+- Set GOOGLE_ANALYTICS to None in settings
+
+3.9.1
+~~~~~
+
+- Fix the issue with switching between child languages
+- Fix the issue with allowing articles to exist in multiple sections
+
+3.9.0
+~~~~~
+
+- Update user permissions
 
 3.8.3
 ~~~~~